refactor(payment): create a event types function
