+++ src/components/Payment/TicketTypeContainer.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export default styled.section`
+  h2 {
+    font-size: 20px;
+    color: #8e8e8e;
+  }
+  > div {
+    display: flex;
+    gap: 24px;
+    margin-top: 17px;
+  }
+`;
